Struct sixty_two_days_of_rust::ThreadPool

pub struct ThreadPool { /* fields omitted */ }

Create the Thread Pool and Storing Threads. It will create a channel and hold on to the sending side.

Methods

impl ThreadPool

fn new(size: usize) -> ThreadPool

Create a new ThreadPool. The size parameter is the number of thread in pool. The ThreadPool::new use for loop counter to generate an id, creade a new Worker with that id, and store the worter in the vertor. Note that the 'new' function will panic if the size is zero.

Examples

Create a thread pool.

use fifty_nine_days_of_rust::ThreadPool;
let pool = ThreadPool::new(0);     // Panic!!!
let worker = ThreadPool::new(5);   // Created successfully.

fn execute<F>(&self, f: F) where
F: FnOnce() + Send + 'static,